remove benchmarks of other maps, separate flurry benches
