+++ Cargo.toml
@@ -28,36 +28,12 @@ num_cpus = "1.12.0"
 rand = "0.7"
 rayon = "1.3"
 criterion = "0.3"
-chashmap = "2.2"
-dashmap = "3.2"
-cht = "0.1"
-contrie = "0.1"
-hashbrown = "0.6"
 
 [[bench]]
-name = "flurry" 
+name = "flurry_dashmap" 
 harness = false
 
 [[bench]]
-name = "chashmap"
+name = "flurry_hashbrown" 
 harness = false
 
-[[bench]]
-name = "dashmap"
-harness = false
-
-[[bench]]
-name = "cht"
-harness = false
-
-[[bench]]
-name = "contrie"
-harness = false
-
-[[bench]]
-name = "rwlock_std"
-harness = false
-
-[[bench]]
-name = "hashbrown"
-harness = false
